Bug: Use the same value type to avoid overflow
The type of stat->len is size_t while the type of nbytes,nlines
and etc is int.

When the value of state->len is assigned to nbytes(lib_mem.c:142),
it has a potential risk to have overflow.

+++ src/lib_mem.c
@@ -128,7 +128,7 @@ tlb_cleanup(iter_t iterations, void* cookie)
 void
 base_initialize(iter_t iterations, void* cookie)
 {
-	int	nwords, nlines, nbytes, npages, nmpages;
+	size_t	nwords, nlines, nbytes, npages, nmpages;
 	size_t *pages;
 	size_t *lines;
 	size_t *words;
